+++ packages/aside-keys/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 24-Feb-2021 (0.0.2-alpha.1)
+
+- Converted to Firebase 9.x API
+
 ## 24-Feb-2021 (0.0.1-alpha.2)
 
 - Dependencies updated. Demo works. First publish to `npm`. :)